[Broadcast] Moving Broadcaster classes to a new Manager namespace for better organization.
