+++ src/Valkyrja/Application/Env.php
@@ -25,8 +25,8 @@ use Valkyrja\Auth\Entity\User as UserEntity;
 use Valkyrja\Auth\SessionAuthenticator;
 use Valkyrja\Auth\Store\Contract\Store;
 use Valkyrja\Auth\Store\OrmStore;
-use Valkyrja\Broadcast\Contract\Broadcaster;
-use Valkyrja\Broadcast\PusherBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\Contract\Broadcaster;
+use Valkyrja\Broadcast\Broadcaster\PusherBroadcaster;
 use Valkyrja\Cache\Manager\Contract\Cache;
 use Valkyrja\Cache\Manager\RedisCache;
 use Valkyrja\Cli\Middleware\Contract\CommandDispatchedMiddleware;
