+++ src/Valkyrja/Broadcast/Broadcaster/LogBroadcaster.php
@@ -11,11 +11,11 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Broadcast;
+namespace Valkyrja\Broadcast\Broadcaster;
 
 use JsonException;
 use Override;
-use Valkyrja\Broadcast\Contract\Broadcaster as Contract;
+use Valkyrja\Broadcast\Broadcaster\Contract\Broadcaster as Contract;
 use Valkyrja\Broadcast\Data\Contract\Message;
 use Valkyrja\Log\Logger\Contract\Logger;
 use Valkyrja\Type\BuiltIn\Support\Arr;
