+++ src/Valkyrja/Broadcast/Provider/ServiceProvider.php
@@ -17,11 +17,11 @@ use Override;
 use Pusher\Pusher;
 use Pusher\PusherException;
 use Valkyrja\Application\Env;
-use Valkyrja\Broadcast\Contract\Broadcaster;
-use Valkyrja\Broadcast\CryptPusherBroadcaster;
-use Valkyrja\Broadcast\LogBroadcaster;
-use Valkyrja\Broadcast\NullBroadcaster;
-use Valkyrja\Broadcast\PusherBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\Contract\Broadcaster;
+use Valkyrja\Broadcast\Broadcaster\CryptPusherBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\LogBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\NullBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\PusherBroadcaster;
 use Valkyrja\Container\Contract\Container;
 use Valkyrja\Container\Support\Provider;
 use Valkyrja\Crypt\Manager\Contract\Crypt;
