+++ tests/Unit/Broadcast/Provider/ServiceProviderTest.php
@@ -16,12 +16,12 @@ namespace Valkyrja\Tests\Unit\Broadcast\Provider;
 use PHPUnit\Framework\MockObject\Exception;
 use Pusher\Pusher;
 use Pusher\PusherException;
-use Valkyrja\Broadcast\Contract\Broadcaster as Contract;
-use Valkyrja\Broadcast\CryptPusherBroadcaster;
-use Valkyrja\Broadcast\LogBroadcaster;
-use Valkyrja\Broadcast\NullBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\Contract\Broadcaster as Contract;
+use Valkyrja\Broadcast\Broadcaster\CryptPusherBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\LogBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\NullBroadcaster;
+use Valkyrja\Broadcast\Broadcaster\PusherBroadcaster;
 use Valkyrja\Broadcast\Provider\ServiceProvider;
-use Valkyrja\Broadcast\PusherBroadcaster;
 use Valkyrja\Crypt\Manager\Contract\Crypt;
 use Valkyrja\Log\Logger\Contract\Logger;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
